refactor based on peer review
